Move to a custom set of zsh configuration files
This drops oh-my-zsh in favor of a set of custom zsh configuration files
that perform all my initializtion. The idea was stolen entirely from:

https://github.com/xero/dotfiles

I considered using Antigen or something like that but I really don't
want any configuration frameworks. With this setup my time to new shell
on my blazingly fast 2017 iMac is less than a second. Compare that with
10s for my oh-my-zsh configuration to load. That's pretty excellent.

It's not 100% perfect yet but this is functional and I can work from
here to refine things.