Working antidote zsh configuration
Bootstrapping off the zdotdir project I was able to get a pretty decent
rework of my zsh environment going using antidote as the plugin manager.
The whole approach has greatly simplified my reliance on plugins. I've
moved a lot of the stuff into zshrc.d/*.zsh files now and done away with
the over use of the _has function -- I just load everything now as much
as it makes sense to do so.

TBD if this works with my work laptop where I have a greater reliance on
oh-my-zsh plugins, but on my home iMac it's pretty fly.