Dump antidote
And all the plugins

I've moved the zephyr/confd like logic in to .zshrc and verified that my
zsh environment is functional for now.

I'll look in to pulling in to my zshrc.d some of the stuff from the
plugins I liked. But generally I don't need much of what I had and this
simplification is good.

The powerlevel10k prompt now loads correctly for me on new shell
creation, always.