Trying to fix the annoying thing with blink
HOLY HELL! I think this fixed it. The thing where hitting enter at the
end of a line caused blink to select the top of the autocompletion list
is now NO MORE!

Thank you AstroVim for showing me the way here.

I might not even need the keymaps, though they make sense to my brain. I
think it was just the bit in the `list` section that fixed this.
Specifically the `preselect = false` part.