Bind C-g to address nested tmux
I know this is an anti-pattern but I live in tmux on my terminal and
having a nested tmux on my remote homelab box for persistence and long
running commands is hella convenient.

C-a was awkward. C-g feels less so. Though I've already mixed it up with
`C-b g` which I use to bring up the floating terminal.