Try CodeCompanion
The documentation and UI polish for codecompanion are just miles beyond
what avante has. Even if avante has more features, which to my untrained
eye doesn't seem to be the case, I'll take a well-organized community
over features any day.

I was using it to write the configuration and customize it once I got it
bootstrapped and it was doing a decent job. The Actions palette is
really excellent.