Make tmux popup windows git-dir aware
Now ctrl-b,g makes the popup window in the root of the git project even
if you're deep in the tree of the project. It also names the window
after the root of the project. This means, no matter where you are in
the git project, you work with the same popup window all the time.